backend/requirements.txt
fastapi
uvicorn
SQLAlchemy>=2.0
python-dotenv>=1.0
httpx>=0.24
google-auth>=2.20
